recast intends visitors to be shared
* extract visitor to own file
* use shared visitor
