+++ lib/index.js
@@ -5,93 +5,7 @@ var through = require('through');
 var b = recast.types.builders;
 var n = recast.types.namedTypes;
 
-function Visitor() {
-  types.PathVisitor.call(this);
-}
-
-Visitor.prototype = Object.create(types.PathVisitor.prototype);
-Visitor.prototype.constructor = Visitor;
-
-Visitor.prototype.visitProperty = function(path) {
-  var expr = path.node;
-
-  if (expr.method && n.FunctionExpression.check(expr.value)) {
-    expr.method = false;
-
-    if (refersTo(expr.value, expr.key)) {
-      var fnArg = astUtil.uniqueIdentifier(path.scope, expr.key.name);
-      var wrapperFn = astUtil.uniqueIdentifier(path.scope, 'wrapper');
-
-      expr.value = b.callExpression(
-        b.functionExpression(
-          null,
-          [fnArg],
-          b.blockStatement([
-            b.variableDeclaration(
-              'var',
-              [
-                b.variableDeclarator(
-                  wrapperFn,
-                  b.functionExpression(
-                    expr.key,
-                    [],
-                    b.blockStatement([
-                      b.returnStatement(
-                        b.callExpression(
-                          b.memberExpression(
-                            fnArg,
-                            b.identifier('apply'),
-                            false
-                          ),
-                          [
-                            b.thisExpression(),
-                            b.identifier('arguments')
-                          ]
-                        )
-                      )
-                    ])
-                  )
-                )
-              ]
-            ),
-            b.expressionStatement(
-              b.assignmentExpression(
-                '=',
-                b.memberExpression(
-                  wrapperFn,
-                  b.identifier('toString'),
-                  false
-                ),
-                b.functionExpression(
-                  null,
-                  [],
-                  b.blockStatement([
-                    b.returnStatement(
-                      b.callExpression(
-                        b.memberExpression(
-                          fnArg,
-                          b.identifier('toString'),
-                          false
-                        ),
-                        []
-                      )
-                    )
-                  ])
-                )
-              )
-            ),
-            b.returnStatement(wrapperFn)
-          ])
-        ),
-        [expr.value]
-      );
-    } else {
-      expr.value.id = expr.key;
-    }
-  }
-
-  return expr;
-};
+var Visitor = require('./visitor');
 
 function refersTo(fn, identifier) {
   var result = false;
@@ -110,7 +24,7 @@ function refersTo(fn, identifier) {
 }
 
 function transform(ast) {
-  recast.visit(ast, new Visitor());
+  recast.visit(ast, Visitor.visitor);
   return ast;
 }
